refactor: add a hook usePinnedModels for pinned models
